fix: ensure that async function execution is waited for
This commit makes the benchmark `onComplete` function async and waits for any
potential return values from the benchmarked function. If an async function is
under test, then the `await` keyword ensures that the value of the returned
`Promise` is logged. If the function under test is not async, the `async`
keyword essentially has no effect except under the covers where the function's
return value is converted to a resolved `Promise` and the resolved value is logged.